refactor(runtime): move model_catalog / audit / plugin_runtime inline tests into their own files
Three more files cleared of inline test blocks, satisfying #3710's
< 2,000 LOC file-size cap for each:

- `model_catalog.rs`: 3,255 -> 1,484 LOC (-1,771). Tests block was in
  the middle of the file (L1280-3050); the post-tests
  `ProbeResult` struct and `probe_api_key` async fn that the rest of
  the runtime calls into stay in place at the tail of the file.
- `audit.rs`: 2,556 -> 1,103 LOC (-1,453).
- `plugin_runtime.rs`: 2,262 -> 1,912 LOC (-350).

Each parent file declares `#[cfg(test)] mod tests;` at the end so the
suite stays path-equivalent (`<module>::tests::*` runs the same
tests). Dedent + de-nest pulls the previously-inside-mod-tests
`#[test]` fns up one level so test names stay flat and the
`super::*` import resolves to the parent module
(not `<module>::tests::tests`).

The first attempt for `model_catalog` used `awk '/^#\[cfg\(test\)\]/{print NR; exit}'`
plus a "tests run to EOF" assumption, which silently dropped
`ProbeResult` + `probe_api_key`. The corrected helper does brace-
counting from `mod tests {` to find the matching `}`, so any
post-tests code at the tail of the file (model_catalog's case) is
preserved untouched. Bash-side helper sketch:

    awk '... depth tracking on {/} after `mod tests {` ...'

cargo check --tests + cargo clippy --lib both clean.

Refs #3710
